Refactor and clean up the code
